Fix circular-dep TDZ crash and picker alias mismatch
Move coerceBrokenShape() and the broken-shape set into
vertexStylesTransform.ts so the import from storageAtoms is type-only
into graphStyles.ts, breaking the runtime cycle that caused a
ReferenceError during initialization.

Change the coercion target from "round-rectangle" to "roundrectangle"
(the value NODE_SHAPE uses for the picker) so the style dialog displays
"Round Rectangle" instead of a blank field for coerced entries.
