+++ packages/graph-explorer/src/core/StateProvider/graphStyles.ts
@@ -45,31 +45,6 @@ export const SHAPE_STYLES = [
 ] as const;
 export type ShapeStyle = (typeof SHAPE_STYLES)[number];
 
-/**
- * Shapes that cytoscape's round-polygon renderer draws incorrectly at small
- * node size (24px): their corner computation degenerates, rendering as a blob
- * and producing invalid edge endpoints that cause edges to disappear. These are
- * kept in {@link SHAPE_STYLES} so older files still parse, but are coerced to
- * `round-rectangle` at every read boundary.
- */
-const BROKEN_ROUND_POLYGON_SHAPES: ReadonlySet<ShapeStyle> = new Set([
-  "round-triangle",
-  "round-pentagon",
-  "round-hexagon",
-  "round-heptagon",
-  "round-octagon",
-  "round-tag",
-]);
-
-const BROKEN_SHAPE_REPLACEMENT: ShapeStyle = "round-rectangle";
-
-/** Coerces a broken round-polygon shape to `round-rectangle`, passing all others through. */
-export function coerceBrokenShape(shape: ShapeStyle): ShapeStyle {
-  return BROKEN_ROUND_POLYGON_SHAPES.has(shape)
-    ? BROKEN_SHAPE_REPLACEMENT
-    : shape;
-}
-
 export const LINE_STYLES = ["solid", "dashed", "dotted"] as const;
 export type LineStyle = (typeof LINE_STYLES)[number];
